test: Add wrapper for commands that need to be run as root
Many tests use a duplicated snippet of code to check if it is needed to
prefix certain commands with `sudo` and to find out the full path of the
`sudo` executable.

Moving these details to a `bats_sudo` wrapper simplifies the tests and
provides a single point to patch to integrate these tests in environments
where finding the path of `sudo` (or of a `sudo`-like command) is non
trivial, for example inside the Debian testing infrastructure.